IDS-FL-CSE400-HR/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pytorchexample"
version = "1.0.0"
description = "Federated Learning with PyTorch and Flower (Quickstart Example)"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.19.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.5.1+cu121",
    "torchvision==0.20.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "pytorchexample.server_app:app"
clientapp = "pytorchexample.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 5          # BASIC: Single round first
fraction-evaluate = 1.0
fraction-fit = 1.0
local-epochs = 5               # BASIC: Very simple epoch count
learning-rate = 0.001          # BASIC: Standard learning rate
batch-size = 64                # BASIC: Small batch size
min_available_clients = 3
server-device = "cuda"

[tool.flwr.federations]
default = "local-simulation-gpu"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 5     # INCREASED: 2→3 for more federated diversity

[tool.flwr.federations.local-simulation-gpu]
options.num-supernodes = 5     # INCREASED: 2→3 clients
options.backend.client-resources.num-cpus = 2
options.backend.client-resources.num-gpus = 0.25